release algoritm initial ships
